+++ example/parse_document.dart
@@ -4,7 +4,7 @@ import 'package:http/http.dart' as http;
 void main() async {
   var url = 'https://flutter.dev';
   var response = await http.get(Uri.parse(url));
-  var document = responseToDocument(response);
+  var document = MetadataFetch.responseToDocument(response);
 
   var data = MetadataParser.parse(document);
   print(data);