+++ lib/src/utils/util.dart
@@ -12,19 +12,6 @@ extension GetMethod on Map {
   }
 }
 
-/// Adds getter/setter for the original [Response.request.url]
-extension HttpRequestData on Document {
-  static String? _requestUrl;
-
-  String? get requestUrl {
-    return _requestUrl;
-  }
-
-  set requestUrl(String? newValue) {
-    _requestUrl = newValue;
-  }
-}
-
 String? getDomain(String url) {
   return Uri.parse(url).host.toString().split('.')[0];
 }