+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## 0.4.0
 
 - NNBD (Null Safety) enabled
-- Various API improvements
+- Various API changes
+- `extract()` and `responsetoDocument` are now encapsulated inside the `MetadataFetch` class. This is to avoid global clashes when imported.
+- Removed the `requestURL` extension which served as a static variable attached to `http.Document`, which caused problems.
+- `Metadata.parse()` now accepts the `url` keyword argument. This is used as a fallback url in `Metadata.url` and to resolve relative (non-absolute URL) images.
 
 ## 0.3.4